Add a basic web UI for remote control
Uses elnode to provide a web UI for remote control of the presentation,
providing the following commands:

  org-present-remote-on
  org-present-remote-off

org-present-remote-on prompts the user for the host on which to
listen (typically the public IP address of the machine), and then starts
a web server on that host, on port 8009.

The endpoints are as follows:

  /prev
  /next
  /

All three endpoints render the same web UI, a basic page with Prev and
Next links.  In addition, /prev invokes org-present-prev on the
presentation buffer, and /next invokes org-present-next.

Note that the buffer of the current presentation is stored as
org-present-remote-buffer when org-present-remote-on is called, and all
calls to org-present-next and org-present-prev will be made on that
buffer.

TODO:

 * authentication + authorization (perhaps just password based?)
 * a more informative UI (display current slide, and maybe a tree)