Add org-present-after-navigate-functions abnormal hook
This adds a single abnormal hook (i.e., a hook which takes arguments)
to org-present.

The hook is called after navigating to a new heading (including the
initial heading after org-present is invoked).  It takes two
parameters:

* the name of the presentation file, and
* the text of the heading.

E.g. the following code:

(add-hook 'org-present-after-navigate-functions
          (lambda (title heading)
            (message (concatenate 'string title " - " heading))))

... will update the minibuffer with the title and heading every time
the user navigates to a new heading.