feat(folds): add org-present-startup-folded a la org-startup-folded
This change seems appropriate to have in base `org-present` rather than as a
hook, because it accords with both parts of `org-present`.

 1. It fits with the `org` part because it is equivalent to
 `org-startup-folded`, but for presentation mode.
 2. It fits with the `present` part because it is analogous to the introduction
 of slide items by effects in other presentation programs: i.e., if you do not
 want to show the whole slide at first, but to unfurl it slowly.

We could have driven this feature of the `org-startup-folded` variable itself,
but it's probably making too much of a assumption about users' preferences to do
that. I want to make sure that this kind of change is opt-in for those who want
it and doesn't disturb existing users by suddenly changing behavior on an update.